requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
elasticsearch>=8.11.1
pinecone-client>=3.0.0

# PDF Processing
PyPDF2>=3.0.0
pdf2image>=1.16.3
pytesseract>=0.3.10
Pillow>=10.0.0

# Security
python-jose>=3.3.0
passlib>=1.7.4
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
httpx>=0.26.0

# Logging
loguru>=0.7.2